+++ DLVN-SDK/src/main/java/com/edoctor/dlvn_sdk/webview/SdkWebView.kt
@@ -253,7 +253,10 @@ class SdkWebView(sdk: EdoctorDlvnSdk): DialogFragment() {
                             EdoctorDlvnSdk.debounceWVShortLink = false
 
                             requireActivity().runOnUiThread {
-                                requestPostNotificationPermission()
+//                                requestPostNotificationPermission()
+                                if (!EdoctorDlvnSdk.dlvnAccessToken.isNullOrEmpty()) {
+                                    requestCameraAndMicrophonePermissionForVideoCall()
+                                }
                             }
                         }
                         super.onPageFinished(view, url)
@@ -362,16 +365,15 @@ class SdkWebView(sdk: EdoctorDlvnSdk): DialogFragment() {
 
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
         return object : Dialog(requireActivity(), theme) {
+            @Deprecated("Deprecated in Java")
             override fun onBackPressed() {
-                if (myWebView.url == domain) {
-                    selfClose()
-                } else if (myWebView.canGoBack()) {
-                    myWebView.goBack()
-//                    if (myWebView.url?.contains("/phong-tu-van") == false) {
-//                        myWebView.goBack()
-//                    }
-                } else {
-                    dismiss()
+                requireActivity().runOnUiThread {
+                    if (myWebView.url == domain) {
+                        selfClose()
+                    }
+                    if (!myWebView.canGoBack()) {
+                        selfClose()
+                    }
                 }
             }
         }